K8S-1858: Allow Recoverable Errors
3rd party tolling can create resources in any order it wants to.  If
they created the cluster before any secrets, shame on them, but the DAC
would reject this as a misconfiguration.  Instead, we should be
eventually consistent, and just error until the situation fixes itself.
While we lose the ability to synchronously tell users off, we can still
do so with conditions, which is actually very useful for testing if
there is a persistent error...

Change-Id: Ie0a5fe26ee1494476023610843e38387c94540d9
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/144628
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>