K8S-1982: Handle Errors in Initialisation
I saw a user example the other day where they configured an illegal
password.  We initialized CBS with it, it decided it was illegal, and we
errored.  Problem was, we'd assume the member was initalized after that
point, and try to talk to it with a busted password.  What we need to do
is if initialization fails, blow the pod away and try again until the
condition that caused the error is fixed.  While we now have password
validation, that is still requisite on the DAC running.

Change-Id: I3f608f9c296d56e933e2fbcec1942bc4c36c49c4
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/145410
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>