K8S-2000: Audit configuration and log streaming
Audit configuration is now handled by the CRD.

The operator will enable/disable auditing via the inconsistent REST API once the pods are up. The various settings available can all be configured and are passed through, the only one that is enforced is log path being set to the default.

No logs are streamed from the current server images. This change allows for a logging sidecar using FluentBit to pick up logs via the shared persistent volume.
The default configuration (if optionally enabled) will stream the audit log to stdout only but this can be extended via a custom Secret with whatever configuration the user wants that works with FluentBit.

This only works if we have a PVC to share the logs with. It could be extended to ephemeral volumes but this needs wider agreement on their use.

Separately there is an issue with the audit logs being rotated but never cleaned up so potentially filling the volume.
This changes also includes a sub-optimal solution (hack) to remove rotated audit log files using busybox sidecar - it also logs what it removes.
The audit log issue really should be resolved by the server.

Currently the following logs can be parsed with the default config:
* audit.log
* babysitter.log
* couchdb.log
* indexer.log
* memcached.log.000000.txt

Only the audit.log is actually streamed though by default - the others could be but are not yet. This change is specifically to cover audit.log only.

Change-Id: I3d98702f4b77ee4fbcce18650029f9adc435031a
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/145825
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Reviewed-by: Eric Schneider <eric.schneider@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>