K8S-2003: Retry Security Setting on Init
We expect server to do an orderly shutdown when we update the server
settings i.e. respond before restarting the network layer.  Alas it
doesn't always indicating that it just hadouken's everything giving no
guarantees to clients, and now we have to handle race conditions.
Instead of treating EOF as "special", we don't really know what's
happened, and this can lead to inconsistency, so retry until success.

Change-Id: I73e1c0e6d0b4c3c94003657fdd53222ee42aa79f
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/146103
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>