K8S-1845: Make Cron use UTC
The cron scheduler is not aware of timezones, and uses the Kubernetes
controller as its local time reference.  Prateek with his +06:30 (or
whatever) timezone doesn't play nicely with all of our clusters (that
appear to use UTC, I've not _seen_ anything to the contrary, and you'd
need to manually set it, which we don't).  Rather than try and derive
the timezone on the Kubernetes cluster, which is pretty much impossible,
just normalize the relative cron times into UTC first.

Change-Id: I9bcfb299bd107d76fa9ae1e679232471aeaaf2ef
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/147172
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Roo Thorp <roo.thorp@couchbase.com>
Reviewed-by: Prateek Kumar <prateekkumar276@gmail.com>