K8S-2047: Refactor E2E utilities
There is a lot of "create cluster with feature X", and "create cluster
with feature X and Y".  This is repetition and leads to bugs.  What we
really want to do is have a single method for enabling X and Y, then
chaining them together.  As a result, creating clusters take on the
builder patten and life is harmonious.

Change-Id: I476b463fd92b828103578316303e2b3e0f4c31a6
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/147359
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Roo Thorp <roo.thorp@couchbase.com>