K8S-2049: Apply API Check to API Operations
The installer tool checks the API version to ensure the customer knows
when they are using an unsupported Kubernetes version (and also denies
them the ability to install on something too old).  This was applied
across the board, which means you cannot generate manifests offline.
Relax the API checks so they are only applied to online operations such
as create and delete.

Change-Id: I96406a37857f6e878d4ea53f95075045f99b262e
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/147435
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>