K8S-1817: FTS documentation and validation updates
Recommendations updated to suggest using a PV for FTS but this also then requires knock-on updates to validation.
The search service reuses the index mount so the DAC now allows that rather than just for the index service.

Tests failed on previous version to confirm validation was wrong, and now pass.

Change-Id: Ife4b8390f392fdae84acb97e85e99704b7a63e59
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/147743
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Eric Schneider <eric.schneider@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>