K8S-2052: Beware Pod Filtering
All pods we create are labelled with our app and the cluster name.  This
is used to drive the fitering for the shared informer cache.  Sadly it
means if you are not careful, then you can end up listing backup pods
too, as was the case here -- the operator pre-run upgrade stuff was
trying to "upgrade" metadata on a backup pod, and failed because there
was no metadata.  Do a quick grep and replace all Pods.List() sites with
the filtered version.  As a follow up, we should see if the Operator
needs to see backup pods at all (it should just be concerned with jobs
and cronjobs...) and we can nip this in the bud by filtering at source.

Change-Id: I051d62d0cf5ed0e6b44341d78d2be36b0f7feec0
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/147801
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>