K8S-1080: redirection of logs to stdout
Standard log output to support various tooling in kubernetes with no changes required to applications or how they log things.
A custom image based on Fluent Bit is provided with full configuration included. This supports the following:
1. Dynamic config reload - change /fluent-bit/config contents and it'll restart Fluent Bit.
2. Rebalance file pre-processing - copy to temporary and add a newline plus additional timestamp from filename.
3. LUA support for redaction.

The container image can be run standalone with the full Couchbase configuration but is intended to be used with the CAO.
The operator will specify the configuration to use but reuse the parser definitions, i.e. not re-specify a copy of them.

Change-Id: I1e95765bf791925b0b21606c40d83e245f52ff60
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/145665
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Eric Schneider <eric.schneider@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>