K8S-2060: Fix LDAP Authorization
Two fixes here, first a bool is a bool, not an integer, which is the
problem the paying customer is stuck with.  Second, I note that the
logic doesn't allow you to explcitly disable nested groups... now I may
be wrong, but the way NS server usually works is it will only take
action and modify fields that are actually passed to it.  By always
sending the value it's guaranteed to be either true or false (when
relevant) and we let server handle the idemopotency.

Change-Id: I6e82172246e7fbb5e97c164f592eb78440c804e3
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/148161
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>