K8S-2078: GKE Autopilot CI
Because why manage something when someone can do it for you.  Also,
dynamic provisioning of the cluster means -- hopefully -- no scheduling
errors and we can go as wide as we want, because it's all ephemeral!

The bulk of this is just making the framework deploy stuff in the
correct way for this kind of environment, and also refactoring the
skipping stuff, which is well out of date or only relevant to
unsupported Couchbase versions.

Change-Id: Ib99e8a4f17c699ba01807bf9cfdfc19f61fb15f9
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/148052
Reviewed-by: Roo Thorp <roo.thorp@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>