K8S-2071: Deadlock Holiday
I don't like cricket, oh no, I love it!  What I don't like is deadlocks.
In this situation the Operator was deleted after pod creation, but
before it could be initialized.  The short version is that server
doesn't work until it's been configured, so on a restart, the status
checks we do fail, and we get stuck forever.  We need to annotate pods
to say they are initialized, and be allowed to slay them if not.  For
now I'm doing this absolutely when necessary, and only for versions that
are known to have this annotation.

Change-Id: I4840cc6a2c2a164a91bb58a6750d2c29e044996c
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/148696
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>