K8S-2079: Resolve racy audit test case
The TestNoLogOrAuditConfig seems to have a race in the formal test cluster when waiting for the reconcile event.
Log analysis show things are fine, the other test case using async WaitForClusterEvent does not directly update the audit REST API.
As we only expect a single event for this test case, switch over to the less racy MustObserverClusterEvent instead.

Change-Id: Ic6698d9b20183afd85cf4c1fabe545a58a113286
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/148811
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>