K8S-2080: Preserve Pod Version Annotation
The version defines the labels/annotations that are attached to
resources, e.g. pod.couchbase.com/initialized, and those can only be
reliably set at the time the pod was created.  We could guess and patch
it in in the operator upgrade code, but that's a bit smelly and also a
pure guess!

Change-Id: I08abf5eafc187906745b98d68c435207dbd6c192
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/148813
Reviewed-by: Pat <patrick.stephens@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>