K8S-1270: Autoscaling Backup PVCs
Horrid code, waste of time, why can't people just monitor things!!  Rant
complete... The size of the volume is the larger of the requested size,
or the actual size (thus handling the fact that you can't emsmallen
volumes).  When enabled, the autoscaler will trigger a resize, when N%
of the storage is left, and increase by M% of the current size to
provide more space.

Change-Id: I2d3e538ddd1e4a1990719755e08df2233969cbdf
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/148848
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>