K8S-2143: Simplify Validation
That one big huge constraints functions is an arse, as it accumulates
errors meaning lots of control flow to avoid going places you shouldn't.
Break this down into a set of feature specific validators.  Also, we're
seeing more platforms ban the use of mutation, so kill this off, this
has the side benefit of making us less complacent when people decide not
to use the DAC.

Change-Id: I9e7ae6845385ea5e0afc29cce401a1c58e181fa8
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/154781
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>