K8S-2143: Simplify Pod and PVC Management
As with all things, simple code comes down to the correct data
structure.  When processing PVCs we need to first determine what's
required for a member, this takes the form of a mapping from member to
mount name and the claim involved.  We then generate required volumes,
process them so we know what action to take, then generate any volume
mounts required to attach them to a container.  While I'm still not
particularly happy with this area, it's simpler at least.

Change-Id: If5deb707519a070afc30f78fd4fc2d89989ff81a
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/154782
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>