K8S-2140: Make SGW Docs More Realistic
People just follow the docs like sheep, we need to point out that you
can control the number of threads in use and how they are connected with
resource scheduling.  It may also be a good idea to add some ports in so
users don't start moaning when they use istio.

Change-Id: Ia057aa1f5e0db89794798bf3cc6f236dbd4e25ca
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/152323
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>