K8S-2204: Update image digests used by OLM
Here we go.  So the alternative here is taking the
digest version and mounting it as a volume via downard api
but there isn't an advantage in doing this since we still
need to decode what the actual semver is of the digest.
And I don't think operator should be quering Red Hat container catalog
for this either.

Ref: https://docs.openshift.com/container-platform/4.6/operators/operator_sdk/osdk-generating-csvs.html#olm-enabling-operator-for-restricted-network_osdk-generating-csvs

Change-Id: I18513762ac528c937a5bf51f0821f76e95361b62
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/154088
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>