K8S-2288: Check Pod Uninitialized Before Kill
Server refuses to work properly until it's had a few calls made to it,
therefore we need to maintain state telling us whether the pod is
initialized (should work) or not (will reject API requests).  In not
initialized, then we can safely delete it, knowing that it's broken as
we have no idea where to restart the initialization process from.  Old
2.0 and 2.1 resources are ignored as we know they can't have the
annotation (at least until topology is updated to upgrade the pod).

First problem is we inject this annotation outside of the usual pod
creation path, so it needs to be preserved as a "3rd party" annotation
(handled by Pat).  Second is we were totally ignoring this annotation
when blowing away pods.

Change-Id: I2880060994c0f459d428a8503fc59b9ac651ff43
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/156746
Reviewed-by: Pat <patrick.stephens@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>