K8S-2293: Update to ensure NodePorts are correct for views
The 8092 port is used by views and should be accessible for `data` services rather than `index` services.
Unfortunately we only test with data+index together so created a new topology for splitting them to retest with.

Turns out 8092 is also used by XDCR and should be available to all, as per the other data service ports.

Added a feature tag for the networking tests as well.
Updated the test to check that the node ports match what we've told Couchbase Server API.
Also verified with the test case supplied by the user.

Change-Id: Ie9037cd691294bdfdc18345563372d143d02a66b
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/157098
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Roo Thorp <roo.thorp@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>