K8S-2330: Scopes & Collections (Documentation)
Update the docs version to be beta, and the image defaults to be beta1.
Reorganize the navigation bar for concepts as it's getting too
cluttered.  Resets the release notes for the 2.3 branch.  Finally add in
the scopes and collections concepts documentation, with lots of pretty
and easy to understand pictures.

Change-Id: Ibe0dee5f1610e14644d7ecb45ebc2223df2d9234
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/159004
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>