K8S-2441: Cull CRD Cruft
We allow too much to be configurable, which shouldn't be there to tempt
users in the first place, but also makes the CRD huge, and thus breaks
kubectl apply (~640k max annotation size).  This starts at 666K for the
whole CRD file, and reduces it to 360K, thus buying us more time!

Change-Id: I26119409d2a22f31483f1b438973411a82496d30
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/162702
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>