K8S-2549: 7.1 TLS Fixes
Two fixes, first is caching of TLS data, so we can perform "atomic"
updates across the whole reconcile loop without data changing under our
feet, resulting in a more reliable and predicatable outcome.  The second
is more horrid.  So when upgrading to TLS, we need to install the CA
certificate first (perhaps a legacy problem??), which means that the
shadow secret needs to be in place on all 7.1+ pods, even if it's not
populated, so we can insert a certificate when the time comes.

Change-Id: I0c7bbb58e24745fac3a3c08c113804143a69983d
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/167199
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>