K8S-2563: Network Settings Fix
When we create intended network settings, we add bits based on server
version, nothing we can do about that.  However when reading the
settings back out to compare, there is no analogue call because the
Couchbase API is utterly terrible, and thus we need to apply the same
logic.  We wouldn't have to if it was designed properly because the
field wouldn't exist in the JSON and implictly wouldn't get populated
when we unmarshal.  Anyway we were comparing nil with a pointer to
false, which is a perpetual mismatch, and spammed the API with updates.
I hope someone reads these, and pays attention... we'd waste a lot less
time and money, and have fewer bugs if it was designed by a professonal
rather than ad hoc.

Change-Id: I49942f31ac646f5723966b36aa4599fb37ef1207
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/167648
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Roo Thorp <roo.thorp@couchbase.com>