K8S-2287: CA Pool
Server 7.1 supports multiple CAs, so you server can have one PKI and
clients another, and you can do snazzy things like TLS rotation in a
piecemeal fashion.  This decouples CAs from sever certs, and treats them
as a separate entity.  The only real hack is to pick the correct CA that
is used by the client to verify the server connection.  While we could
just pass in the whole pool, it would need saving in the persistent
secret, which is troublesome to say the least.  One for another day.

Change-Id: Ieaeae460679aacb3e4a6d30ed4cb31148e1a8de6
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/168429
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>