K8S-2478: Revamp the Make System
The primary goal here is to have binaries available so customers can
build their own images from our downloads, when they are unable to
access an image registry... yes, sounds stupid, and it is.  The other
thing we've done is ensure all builds happen in-container, and this is
controlled exclusively by us, and not the build team.  Also we've made
adding a target platform really easy, like one line easy for tools and
images.

Change-Id: If4b5e1aa4067a1acf26dc0a9942ad4b1deda0c28
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/171173
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>