K8S-2654: Fix Potential Nil Pointer
Customer hit nil pointer dereference when upgrading to 2.2.2.  I seem to
have already handled half of it in K8S-2473, but there is another to
catch.  Story goes operator is updated, new CRD values aren't defaulted
because there hasn't been a write yet, and we crash.  Add in some sane
defaults for this case.

Change-Id: I8b33d4cfd76dad36f5a5aa83810eecbd1af54f95
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/171672
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: alex emery <alex.emery@couchbase.com>