K8S-2478: Revamp Makefile - More Fixes
Intesting one this, by driving more stuff from the same source of truth,
we discovered that build "expects" (in the sense a registry needs to be
manually created) the admission controller to be called just that, not
couchbase-operator-admission as it has been since the dawn of creation.
The build's in the right here, and we should rename our binaries to have
parity with that.

Change-Id: I9f1cebac583c8691018e1ff9b8d41ef895f1ea96
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/171912
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>