K8S-2665: Upgrade Recovery Fix
If an original pod goes down during upgrade, then it will be recovered
with a new image version, which sadly is incompatible at the data
structure level, so we end up evidently not only in a death loop, but
also corrupt the data in the process.  Add PVC metadata to remember the
exact image in use, so it can be used for recovery.  Also, plug the
testing hole that allowed this to happen in the first place.

Change-Id: I973ebb8df4ac132c2142401df7d8e91f24ce842f
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/172197
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>