K8S-2679: Fix Blasted Foreign Nodes
Not once in 5 years has a customer ever done this, yet for 5 years we
have been supporting this annoyingly constantly broken test :rage: This
simply treats the pod as having a missing server class and it gets nuked
that way.  I've changed the rebalance logic to prevent a race, and also
the test to expect the new behaviour.

Change-Id: Ic80cb6c915266f08af7c08b96685230ad681034e
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/172438
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Prateek Kumar <prateekkumar276@gmail.com>