K8S-2266: Support for private key passphrase registration
"With Couchbase Server 7.1.0, we allow administrators to use an encrypted private key.
If an attacker could get hold of the private key, they wouldn’t be able to use it without the associated passphrase."

The operator supports passphrase retrieval via script/rest.  The rest solution is very straight foward in that Operator
will call the given URL.  The script solution involves some proxying for security where Operator creates the script
and provides a secure means of it's execution and passphrase retrieval.

Change-Id: I671be8a6d047eec56431b16e54837df81700e44f
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/179551
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Justin Ashworth <justin.ashworth@couchbase.com>