K8S-3034: use pod hostname for requests
Use the pod hostname directly for requests instead
of using the service we create.

Incases where external DNS is used, we end up getting an external IP for
the DNS lookup of the service name.
If there's an issue with the external address we can't actually talk to the pod... despite the fact its
internal address is fine.

Change-Id: I761f3f16503829358cbc994912c9508ea46c5a8b
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/192157
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike <mike.stanton@couchbase.com>