K8S-2644: document Helm DAC cert rotation.
For Helm the DAC certificate would expire after a year.
Whilst we've extended it to 10 years, there are old deployments
still affected, so we need to document how to rotate and get out of that
situation.

Since the cert is hard coded in the webhook's configuration, we need to
delete them and create a new secret, which Admission Controller will hot
reload.

Change-Id: If9fab3aef1a870d7f4be50b8519d3722a1034ec2
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/192598
Reviewed-by: Justin Ashworth <justin.ashworth@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>