K8S-3084: Fixed data thread validation
As of Server 7.1, the min Writer/Reader threads has changed to 4 from 1.

This means we need to accept 1 as a min value, but then check if we're
running 7.1 or higher.

Since go's default zero value for ints is 0, this caused validation to
fail if a user specified 1 thread and not the other.
This would cause the struct to be instantiated with the value provided,
but also default the other to 0.

So all thread variables are now pointers not ints so we can nil check
them first.

Change-Id: I70ef6c768ab51cd7ed19cc8a2bd360f209720ae7
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/192743
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Justin Ashworth <justin.ashworth@couchbase.com>