K8S-3111: improved dns propagation logging
Reverts K8S-3034 and instead improves logging around dns propagation.
It was unclear when DNS issues were met that the cause was an explicit
wait for DNS propagation, as opposed to Operator trying to use the
external DNS name for communication. This change adds hostname to log
messages for better clarity.

Change-Id: I472fd0962ca6c1c83e7f8de52669c613095a9a5b
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/194141
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Justin Ashworth <justin.ashworth@couchbase.com>