Fixing the yaml, context and env precedence
Action config fields can be set in context and used by subsequent actions without redefining it.
Action config fields can also be fetched from env.

When fields are present in yaml, context and env the yaml takes first precedence. If the yaml is empty, context takes
precedence over env.

The first time a field is encountered with a context tag, it will be set to context as well(post field value fetch from yaml/env).
If the field is overriden in subsequent actions(from yaml), the context field will not be overriden. The onus falls onto the action to set it.

If context value is nil and value already present in field (from yaml), the context will be set with fieldValue

If the context value is not nil and field value is nil, then the field value will be set from context

If the context value is not nil and field value is not nil, then the field value takes precedence.
Also this does not override the context value. The previous context value remains.
The override of the context value has to be taken care of by the action
What this implies is that the context value will be set the first time a fieldValue is encountered.
Further, the context value will not be overriden by the new fieldValues and that has to be taken care by the action.

Change-Id: I3b16265cfdb3ad2543d94409dd460c675b0034d3
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/215988
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Raghav S K <raghav.sk@couchbase.com>
Reviewed-by: Raghav S K <raghav.sk@couchbase.com>
Reviewed-by: Aryaan Bhaskar <aryaan.bhaskar@couchbase.com>