Changing action interface
The action interface had Checks() and Do() as part of its functionality.
It is replaced with the following
1. Do()
2. CheckConfig()
3. RunValidators()

All config param checks are added to CheckConfig()
CheckConfig() is run first immediately after it is reconciled.
RunValidators() are run pre and post Do()

Do() and Checks() were also taking the config interface as a param pass by value,
it was never used. Removed the unnecessary params.

Change-Id: I6b453a9c6cbebc9bb46e208a3b9bc102bc69f411
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/221594
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Aryaan Bhaskar <aryaan.bhaskar@couchbase.com>