Recursively reconciling structs in ActionConfig
Previously only the ActionConfig is reconciled. If it holds any struct or ptr to struct, those are not reconciled.
This commit fixes that by calling the ReconcileConfig() recursively when a struct type is detected.
The ctx, env values are populated in the nested structs as well. Also the required tag will also be checked for nested struct fields.

Change-Id: I46d720a0629969cd0619a774db85128dd792f70a
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/222739
Reviewed-by: Raghav S K <raghav.sk@couchbase.com>
Reviewed-by: Aryaan Bhaskar <aryaan.bhaskar@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Raghav S K <raghav.sk@couchbase.com>