Recursively reconciling slices in ActionConfig
Previously only the ActionConfig and nested structs were reconciled. If it holds any slice of structs or slice of ptr to structs, those were not reconciled.
This commit fixes that by calling the ReconcileConfig() recursively when a slice type is detected.
The ctx, env values are populated in the nested lists,structs as well. Also the required tag will also be checked for nested lists,struct fields.

Change-Id: I9b82b5210916345e0e0e45031ec61ca722c8e9db
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/222742
Reviewed-by: Raghav S K <raghav.sk@couchbase.com>
Reviewed-by: Aryaan Bhaskar <aryaan.bhaskar@couchbase.com>
Tested-by: Raghav S K <raghav.sk@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>