K8S-4102: Tolerate larger-than-requested persistent volumes
When a storage class provisions a volume larger than requested, the operator incorrectly assumes a resize is in progress and aborts reconciliation.

This change allows reconciliation to proceed if the actual volume size is greater than or equal to the requested size, fixing the reconciliation loop for both server and backup volumes.

Change-Id: I143939b59c541ecb1f8a95934501a016dc42dcf7
Reviewed-on: https://review.couchbase.org/c/couchbase-operator/+/231565
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <usamah.jassat@couchbase.com>